+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/SettingsStatus.java
@@ -64,6 +64,7 @@ public class SettingsStatus {
     public static boolean inboxFilterEnabled = false;
     public static boolean videoFitEnabled = false;
     public static boolean refreshRateEnabled = false;
+    public static boolean duetStitchEnabled = false;
     public static boolean notificationControlsEnabled = false;
     public static boolean hideSuggestedAccountsEnabled = false;
     public static boolean hideInboxStoriesEnabled = false;
@@ -186,6 +187,10 @@ public class SettingsStatus {
 
     public static void enableExpandActivityList() {
         expandActivityListEnabled = true;
+    }
+
+    public static void enableDuetStitch() {
+        duetStitchEnabled = true;
     }
 
     public static void enableRefreshRate() {